used thunk to fix login
