+++ src/js/actions/CoreActionConsts.js
@@ -30,8 +30,8 @@ module.exports = {
   SET_USER_NAME: "SET_USER_NAME",
   SET_USER_PASSWORD: "SET_USER_PASSWORD",
   TOGGLE_ACOUNT_VIEW_TO_LOGIN: "TOGGLE_ACOUNT_VIEW_TO_LOGIN",
-  LOGIN_USER: "LOGIN_USER",
   LOGOUT_USER: "LOGOUT_USER",
+  RECEIVE_LOGIN: "RECEIVE_LOGIN"
 };
 
 /**
